Gunnar: bugfix for missing fct in RealVector
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1305 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/Hamiltonian/AbstractHamiltonian.cc
@@ -592,7 +592,7 @@ RealVector& AbstractHamiltonian::ConjugateLowLevelMultiply(RealVector& vSource, 
 RealVector& AbstractHamiltonian::ConjugateLowLevelMultiply(RealVector& vSource, RealVector& vDestination, 
 							   int firstComponent, int nbrComponent)
 {
-  vDestination.ClearVectorSegment(firstComponent, nbrComponent);
+  vDestination.ClearVectorSegment((long)firstComponent, (long)nbrComponent);
   return this->ConjugateLowLevelAddMultiply(vSource, vDestination, firstComponent, nbrComponent);
 }
